enhancements in controller, views, and rake tasks

+++ lib/tasks/refresh_srs.rake
@@ -3,72 +3,38 @@ task :refresh_srs => :environment do
   # includes
   require 'open-uri'
   require 'json'
-
-  # importing json confidentials
-  conf_file = File.open("config/confidential.json", "r:utf-8").read
-  conf_json = JSON.parse(conf_file)
-
+   
+    
   puts "Refreshing service request objects in the database.."
   ServiceRequest.all.each do |request|
-    # Get updated Queue information for each SR
-    honed_desarray = []
-    honed_stamparray = []
-    honed_queuearray = []
-    honed_contactarray = []
-    honed_priorityarray = []
-    honed_localearray = []
-    uri = open("#{conf_json["sr_info"]}#{request.number}") do |content|
-      read_content = content.read.force_encoding("ISO-8859-1").encode("utf-8", replace: nil)
-      # last act
-      desarray = read_content.split("DESC_TEXT = ")
-      honed_desarray = desarray[1].split("X_LAST_ACT_COMMENT = ")
-      # There is not always a Last Act Comment, this is to avoid an error in the creation of the object
-      if honed_desarray[1].nil?
-        honed_desarray[1] = ""
-      end
-      # Last Act Stamp
-      stamparray = read_content.split("ACTL_RESP_TS = ")
-      honed_stamparray = stamparray[1].split("<br>")
-      # Locale
-      localearray = read_content.split("ORG = ")
-      honed_localearray = localearray[1].split("<br>")
-      # Queue
-      queuearray = read_content.split("LOGIN = ")
-      honed_queuearray = queuearray[1].split("<br>")
-      # contactvia
-      contactarray = read_content.split("X_RESPOND_VIA = ")
-      honed_contactarray = contactarray[1].split("<br>")
-      # priority
-      priorityarray = read_content.split("SR_SEVERITY = ")
-      honed_priorityarray = priorityarray[1].split("<br>")
-    end
+  
+    # importing json confidentials
+    conf_file = File.open("config/confidential.json", "r:utf-8").read
+    conf_json = JSON.parse(conf_file)
+
+    # pull down SR content for updates. Array to hash in order to take care of html crap. (why couldn't it be json?? (╯°□°)╯︵ ┻━┻)
+    read_content = open("#{$conf_json['sr_info']}#{request.number}", "r:utf-8").read.force_encoding("ISO-8859-1").encode("utf-8", replace: nil)
+    content_array = read_content.split("<HTML>\r\n   ", 2)[1].split("<br>")
+    sr_content = Hash[content_array.map { |item| item.split(" = ", 2) }]
 
     # unless the rosa db still matched the proetus db, update the rosa db to match the proetus db
-    unless ( request.lastact == honed_desarray[1] )
-      puts "Updated #{request.number} last act"
-      request.update_attribute(:lastact, honed_desarray[1])
-    end
-    unless ( request.lastactstamp == honed_stamparray[0] )
-      puts "Updated #{request.number} last act stamp"
-      request.update_attribute(:lastactstamp, honed_stamparray[0])
-    end
-    if request.taken
-      unless ( request.username.name == honed_queuearray[0] )
-        puts "It saw that #{request.number} no longer matches in ownership - Destroying #{request.number}"
-        request.destroy
+  
+    sr_references = Hash[ request.lastact => "X_LAST_ACT_COMMENT", request.lastactstamp => "ACTL_RESP_TS", request.username => "LOGIN", request.priority => "SR_SEVERITY", request.contactvia => "X_RESPOND_VIA", request.locale => "ORG"]
+
+    sr_references.each do |key, value|
+      if (key == request.username)
+        if request.taken && !( key.name == sr_content["#{value}"])
+          puts "It saw that #{request.number} no longer matches in ownership - Destroying #{request.number}"
+          request.destroy
+          puts "#{request.number} destroyed."
+        end
+      else
+        unless ( key == sr_content["#{value}"] )
+          puts "Updating #{request.number} #{key}"
+          request.update_attribute(:key, sr_content["#{value}"])
+          puts "#{key} updated."
+        end
       end
-    end
-    unless ( request.priority == honed_priorityarray[0] )
-      puts "Updated #{request.number} priority"
-      request.update_attribute(:priority, honed_priorityarray[0])
-    end
-    unless ( request.contactvia == honed_contactarray[0] )
-      puts "Updated #{request.number} contact method"
-      request.update_attribute(:contactvia, honed_contactarray[0] )
-    end
-    unless ( request.locale == honed_localearray[0] )
-      puts "Updated #{request.number} locale"
-      request.update_attribute(:locale, honed_localearray[0] )
     end
   end
   puts "Done."
